+++ src/test/java/nextstep/subway/line/infrastructure/InMemoryLineRepository.java
@@ -1,9 +1,9 @@
-package nextstep.subway.line.infra;
+package nextstep.subway.line.infrastructure;
 
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
 import nextstep.subway.line.domain.Section;
-import nextstep.subway.station.infra.InMemoryStationRepository;
+import nextstep.subway.station.infrastructure.InMemoryStationRepository;
 import nextstep.subway.station.domain.Station;
 import org.springframework.test.util.ReflectionTestUtils;
